+++ app/components/ExamPages/3_ScalesNotate.tsx
