+++ app/components/ExamPages/4_TriadsNotate.tsx
