+++ app/components/ExamPages/6_ChordsIdentify.tsx
